requirements.txt
dvc==2.24.0
geemap==0.17.1
joblib==1.1.0
mlflow==1.28.0
numpy==1.23.2
pandas==1.4.4
plotly==5.10.0
python_dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scikit_learn==1.1.3
setuptools==65.4.1
smogn==0.1.2
streamlit==1.12.2
xgboost==1.6.2
PyYAML==6.0
s3fs==2023.12.2
altair<5